feat(lotto) : Input class 구현

+++ src/main/java/lotto/domain/Profit.java
@@ -3,8 +3,8 @@ package lotto.domain;
 import java.util.List;
 
 public class Profit {
-    public double CalProfit(Integer Amount, List<Integer> result){
-        double profit = 0.2;
+    public float CalProfit(Integer Amount, List<Integer> result){
+        float profit = 0.2f;
         return profit;
     }
